Preserve NUL characters in fs::path::string()
Even though NULs are forbidden in the file paths, std::filesystem::path
doesn't discard them and returns them as part of std::[u8]strings that
it returns, so do likewise rather than truncating the string at the
first NUL appearing in it.

Note that this doesn't affect test_path_validation() since the changes
of the previous commit, but would have broken it for C++20 compilers
with char8_t support without them (previously the test was broken only
for the compilers without char8_t support).