Rename 'irc7702.?pp' --> 'gpt7702.?pp'
The implementation to be committed soon is for GPT only, because
 - it constitutes a coherent standalone abstraction; and
 - CVAT, though trivial, is incongruous with GPT.

Chose the name 'gpt7702' for the files and for the class because
 - 'gpt' alone would be ungreppable
 - names such as 'gpt_impl' or 'gpt_class' gain greppability only by
   appending meaningless suffixes
 - although a -'7702' suffix adds no meaning that 'gpt' alone lacks,
   grepping for '7702' finds all 7702 code
 - 'gpt7702' stands out from existing /gpt_.*/ file names, which is
   good: they're ancillary, but this is fundamental
Considered but rejected 'gptest': it's always "G P T", and nobody ever
speaks of a "G P Test"; and that would leave no good name for existing
file 'gpt_test.cpp'.

It's best to rename everything now before actually implementing it.