Assert that guideline premiums use a correct spec amt charge
The spec amt charge is based on the initial DB, reflecting corridor;
but guideline premiums are set before any initial premium is processed.
This raises the possibility that guidelines are incorrect because the
corridor is not reflected. (They'd be lower than necessary, so the
definitional test would still be met.)

That was almost impossible when the statutory rate was four percent:
the maximum initial premium was based on a six-percent GSP rate, and
that rate times the GPT corridor factor was unlikely to exceed unity,
making it unlikely for a contract to be in the corridor on the issue
date. Perhaps that could happen with a substandard contract. Now that
the statutory rate can potentially be zero, or perhaps even less, it
seems best to assert that the spec amt charge used for guidelines is
the same as the one actually deducted.