Eliminate a side-effect guarantee [280]
Expunged this feature, resolving a marked defect.

UL solves generate a complete set of ledger values for each iterand.
The last two iterands are the best upper and lower bounds, to within
the precision specified by the 'decimals' argument. One of those is
selected according to the 'root_bias' argument, each approximately
half the time, so it had been hoped that the ledger values produced by
the last iteration could be used, saving about half an iteration per
solve. However, that hope was in vain, as the AccountValue::Solving
flag has side effects.