Improve instrumentation
When a trace is to be printed, it is useful to know what interpolation
technique was used at each step.

Brent resorts to bisection in two cases, described on pages 50 and 51 of
_Algorithms for Minimization without Derivatives_ (ISBN 0-13-022335-2):
 - "Our main modification of Dekker's algorithm ensures that a bisection
   is done at least once in every 2 log₂(|b-c|/δ) consecutive steps."
 - "When inverse quadratic interpolation is used the interpolating
   parabola cannot be a good approximation to f unless it is single-
   valued between (f,f(b)) and (c,f(c)). Thus, it is natural to accept
   the point i if it lies between b and c, and up to three-quarters of
   the way from b to c: consider the limiting case where the
   interpolating parabola has a vertical tangent at c and f(b) = -f(c).
   Hence, we reject i if 2|p| ≥ 3|mq|."
The trace distinguishes two cases as 'B' and 'b' respectively, which
may correspond to the two cases above (the ALGOL original is obscure).