Remove check for Latin-9 from coding rules test
This test doesn't work with std::regex and is, in any case, not very
useful as what we really want is to check that the files are correctly
encoded in UTF-8.

Such check will be added later, but for now just remove the existing
check, which currently takes ~1% of the execution time.