Use more standard regex syntax in the raw C++11 strings check
The regex added in fec9d6d9b (Establish the One True nonempty
d-char-sequence, 2021-03-08) used Vim "\<" regex construct which
happened to work with boost::regex but is not recognized by ECMAScript
regex flavour used with std::regex.

Replace it with "\b" which works with both boost and std classes and is
also consistent with the use of "\b" in many other places in the
existing code.

Also add a test for this check to test_coding_rules_test.sh to ensure
that this continues to work as expected even after the switch to
std::regex.