Designate release candidate 20210602T1807Z
Effect on speed is curiously variable:
  +/- 10% for the mean, which experience shows to vary greatly; and
  +5% to -3% improvement for the more robust 'least' time.
The speed tests measure a handful of scenarios, most of which involve
solves. If one or another of those solves requires 19 instead of 20
iterations, the overall speed could change by something like 5%. The
number of iterations is a small integer, on the order of ten or twenty,
making the slightest possible change ten or five percent respectively.

It must be emphasized that this does not mean that lmi has actually
become faster in general. No recent change should have produced such an
overall effect. But some recent changes could very well have perturbed
the number of iterations required for some particular solve in the
limited handful measured.