Treat decimal_root return value more flexibly
At present, enum root_validity has only two values:
    {root_is_valid
    ,root_not_bracketed
    };
but it is desirable to add a third, indicating a different mode of
failure, as all successes are similar but failures can have different
causes. Thus, that enumeration wants to be handled in a 'switch'.

Where failure is considered impossible, assert success, instead of
asserting the absence of some type of failure: then the assertion
makes a 'switch' unnecessary.