Diagnose any attempt to find a root in a zero-measure interval
Trying to find a zero of a function in [k,k] is probably a mistake.
Indeed this diagnostic identified a futile unit test, which attempted
to find a zero of f(x)=x^19 in [0,0] to within a tolerance of 1e100,
which had previously been deemed to succeed. Temporarily suppressed
that test.