Refactor
Incrementing a local variable in three locations is better than passing
it by reference to a separate function that increments it in one place.
More crucially, this change ensures that 'n_iter' records the correct
number of iterations even if a trace is not being printed: that's useful
for measuring worst-case behavior and validating an implementation of
Brent's method against his guarantees.