Expunge an unneeded overload
An overload like irr(container) or irr(begin,end) seems obviously
desirable--it's what spreadsheets offer--so why not provide it here?
One reason is that lmi doesn't use it. Instead of
  IRR(vector) -> scalar
illustrations really need IRR columns
  IRR(vector payments, vector benefits) -> vector
where both input vectors are typically positive.

Another reason is that the expunged implementation was flawed. It had
been implemented in terms of a helper that assumed it would receive
(typically positive) payments and benefits separately--unlike the
spreadsheet function, which typically operates on a vector containing
(at least) one negated value. To adapt the helper to that use case,
the last element of a {pmt0, pmt1, ... pmtN, bft} vector could have
been extracted, un-negated, and passed as a distinct argument; instead,
as a nonchalant expedient, a zero had been appended, causing -100% to
be a root in every case. An attempt to repair the expunged overload by
un-negation proved correct, but too ugly--if such an overload is wanted,
it should be implemented in a straightforward way, but that's too much
work for an overload that lmi doesn't even use.