Expunge two unwanted 7702 database entities
No visible effect because neither of these was used.

RiskyInitial7702Db: As '7702.html' [3/1] explains, a contract should be
  "issued at a SA no lower than its corridor DB (¶4/2) after the
   initial premium, dumpin, and any estimated §1035 exchange amount
   are recognized"
Thus, initial payments are recognized, yielding the highest possible
premium limits, which is what producers want when they ask about this.
The question at hand is whether specified amount (SA) can be set lower;
the answer is "no", because any perceived benefit is probably illusory
and certainly not worth risking the contract's tax status.

DB_Irc7702Endowment: There is never any reason not to use the specified
amount as the endowment amount for GLP and GSP. It produces the highest
possible premium limits, and the statute specifically allows it.