Add database entity 'DB_OmitGuarPremSolve'
The NAIC illustration reg requires that a guaranteed premium be
calculated and shown on illustrations in the first year. Single-premium
products generally call for a large initial premium; may restrict
further premium; and may or may not guarantee coverage for the full
term of the contract. For such contracts, compliance officials may deem
it unnecessary to show the usual guaranteed (level) premium.

If the guaranteed premium is not to be shown, setting this entity to
'true' causes it not even to be calculated, which is a significant
speed optimization.