Refactor instrumented reference implementation
This parallels the immediately preceding commit.

Refactoring production and reference implementations separately can
expose defects in the (identical) refactoring.

ALGOL 60 did not perform short-circuit evaluation of logical {OR,AND}.
Evaluating both conditions explicitly is thus truer to the original;
but that doesn't matter because the affected conditions are free of
side effects and the time to evaluate both is negligible.