Rename an enumeration
Soon it will contain more enumerators, to trace numerous paths through
the algorithm. These aren't exactly "decisions" or "activities" on a
flowchart, or "transitions" on a state diagram; rather, they record the
reason why control has arrived at a certain point.

It will be seen that the control structure may be understood as

  bisect:
    COMEFROM dithering_near_root
    COMEFROM secant_out_of_bounds
    COMEFROM parabola_not_single_valued
    COMEFROM guarantee_linear_convergence
  [etc.]

but that would seem jocular, whereas recording the impetus that led to
bisection is plainly serious.