Show variable shifts in optional trace
Operations 'j' and 'k' are shifts of {a,b,c} that maintain invariants:
    force_b_and_c_to_bracket_root    // j
    force_b_to_be_best_approximation // k
and displaying them explicitly makes the trace easier to follow.

For example, the fourth evaluation of the objective function (resulting
from inverse quadratic interpolation) in this trace excerpt:

#eval       a        fa       b         fb       c       fc
  3 L       5  0.609438 3.80896   0.337356     0.5 -1.69315
  4 Q 3.80896  0.337356 2.57754  -0.053165     0.5 -1.69315
  4 j 3.80896  0.337356 2.57754  -0.053165 3.80896 0.337356
  5 L 2.57754 -0.053165 2.74518 0.00984766 3.80896 0.337356

caused b and c to be on the same side of a root (because fb and fc have
the same sign), but Brent's method requires them to bracket a root--so
c's old value is discarded and replaced by a's value, before choosing
the fifth candidate for evaluation. (It might be clearer to suppress the
redundant "4" in "4 j", but the benefit doesn't justify the complexity.)

Brent's method may be viewed as containing a state machine (whose inputs
are internal). The state consists of:
  three abscissae {a,b,c} and their ordinates {fa,fb,fc}
and there are about ten transitions that affect that state. In the
trace, each element of the state is a column, each transition is a row,
and row labels like '4 j' show the evolution.