Store a raw input filename as well as one for monthly traces
Monthly trace files are already saved with a filename based upon the
input case and cell names, with a distinctive extension. Added a
string that's generally the same without that extension, for writing
diagnostic messages that identify the input source.

The name of the trace file is distinct because it needs to be set OAOO.