Elaborate solve traces
It is occasionally useful to enable the "idiosyncrasyT" behavior in
every case, e.g.:

-    if(contains(yare_input_.Comments, "idiosyncrasyT"))
+    if(true || contains(yare_input_.Comments, "idiosyncrasyT"))

when running a set of input files (as for a system test), thus saving
a trace of every solve in a single file. In that case, it is important
to name the input file and describe the solve.

Removed the exclusion of guaranteed-premium solves because they are
as interesting as any other solve.