Don't hard-code sqrt(2)
Calling sqrt() seems preferable. See this IEEE publication:

  https://ieeexplore.ieee.org/document/910638
| The IEEE 754 standard for floating-point arithmetic requires that the
| four arithmetic operations and the square root should be correctly
| rounded.

It's not entirely clear what it means to "require" that something
"should" be done. Presumably a mandative subjunctive is intended:
  requires that...the square root be correctly rounded
  ["should" inserted only because UK English lacks the subjunctive]
because the IEEE 754 standard itself groups squareRoot in section
  "5.4.1 Arithmetic operations"
along with {+,-,×,÷}, whereas another section
  "Table 9.1 — Recommended correctly rounded functions"
lists functions, like reciprocal-sqrt and the usual transcendentals,
for which correct rounding would be ideal but was not originally
required (the authors go on to show that it is actually feasible).