Controvert a published claim
Scherer (_Computational Physics_, page 97) claims:
  "Chandrupatla's method is more efficient than Dekker's or Brent's,
  especially for higher-order roots (Figs. 6.10, 6.11, 6.12)."
As the present investigation shows, that conclusion is based on a tiny
number of atypical examples, whereas a large number of real-world tests
supports the opposite conclusion, that Brent's method is preferable to
Chandrupatla's (and of course Dekker's, as proven in Brent's book).

Translated Scherer's pseudocode for Chandrupatla's and Brent's methods
to C++ to reproduce those examples and evaluate that claim. (Corrected
some but not all defects in the pseudocode for Brent's method, as noted
inline.)

Chandrupatla's algorithm differs in two ways from Brent's. It never uses
linear (secant) interpolation; instead, it uses only inverse quadratic
interpolation (IQI) or bisection. And it imposes a more restrictive
criterion for IQI, requiring that the interpolating parabola truly be
single-valued in the current interval, where Brent instead heuristically
requires that an IQI iterate move no more than three-quarters of the way
from the current best estimate to the contrapoint at which the sign of
the evaluated objective function is opposite.

Results of lmi's system test:

(1) Using Chandrupatla's algorithm:

+#if defined USE_CHANDRUPATLA_ALGORITHM
+// evaluations  max   mean  sample-std-dev
+//     7332      65   18.9       7.13        HEAD (brent)
+//     9149      59   23.6      11.13        this (chandrupatla)

(2) Using Brent's algorithm, but modified to use Chadrupatla's tighter
IQI acceptance criterion:

+//#define     USE_CHANDRUPATLA_IQI_CRITERION
+// evaluations  max   mean  sample-std-dev
+//     7332      65   18.9       7.13        HEAD (brent)
+//     8545      66   22.0      12.90        this (IQI more stringent)

This real-world example involves 388 test cases, each of which finds a
root of a polynomial of degree approximately six hundred, which is
a higher-order root than the textbook's examples test. Brent's method
is clearly superior. Both ways in which it differs from Chandrupatla's:
 - (conditionally) use secant interpolation
 - heuristically accept some IQI iterates that Chandrupatla rejects
contribute to its superiority.

Analysis of the three textbook examples (see unit test):

Fig. 6.10, f(x) = x^-2: Ten evaluations versus eleven is not a big
difference. The graph seems to show Brent's method taking an extra two
iterations to ensure that a root is bracketed; probably that wouldn't
occur with a slighly different tolerance.

Fig. 6.11, f(x) = (x-1)^3: For this equation, bisection is the fastest
of the algorithms depicted. lmi_root(), with an optional parameter to
force bisection always, is shown to take exactly the same number of
evaluations. Uncommenting the tracing code shows that Chandrupatla's
method chooses IQI nine times, otherwise bisection. In the given
[0.0, 1.8] range, bisection ought to take log2(1.8/DBL_EPSILON) = 53
steps, so the graph for bisection seems correct, and Chandrupatla's
method takes 53+9=62 evaluations.

Fig. 6.11, f(x) = x^25: Here, Chandrupatla's method always chooses
bisection, and the graphs of those two methods virtually coincide.

Thus, the evidence presented consists of one test where Chandrupatla's
and Brent's methods perform similarly, and two where bisection is
optimal and Chandrupatla's bias toward bisection is favored. Against
that thin evidence, the real-world system test above shows Brent's
method to be the better choice.