Make root-finding trace more informative
Curiously, though not astonishingly, showing more information in a trace
has the effect of changing the number of function evaluations required
to find a particular root with x87 only. Experimental (not committed)
investigation shows that the cause is a minuscule difference in an early
iterate, which could be expected with an 80- to 64-bit register spill.