Add a null_stream unit test
Using a std::stringstream with badbit set, or a std::fstream with no
associated file (whether or not badbit is set), is fast and simple.
Techniques that set badbit aren't appropriate for code that tests
operations to make sure they succeed, as they will instead fail.
Another problem with all these techniques is that they might not
work: Kühl noted in 1999 that "some implementations have minor bugs
and do not always consider the flags correctly", and there were still
defects many years later--see, e.g., the comment to this article:
  https://stackoverflow.com/a/8244052
that refers to this defect addressed only in 2017:
  https://gcc.gnu.org/bugzilla/show_bug.cgi?id=53984

The null-stream technique, derived from Dietmar Kühl's work, is much
slower because of its construction overhead; but it's as fast as the
others if it's constructed OAOO. (The OAOO approach currently used by
lmi happens to set badbit, which therefore seems not to be a problem
in practice after all.)

The /dev/null technique is too slow.