Don't write solve trace to statusbar
The calculation time reported on the statusbar, for a particular cell
with a solve that takes dozens of iterations, is:
 - 156 msec (previously) with each iteration written to the statusbar
 -  80 msec (now) without writing details of each iteration
so that indicator of progress took as much time as the operation it was
reporting on. Twenty years ago, each iteration was so slow that it was
at least entertaining to watch the iterates dance, but now they'd become
nothing but a costly blur.