Experimental quasi-branch: TOMS 748
This is a quasi-branch:
 - it has no effect outside the code it adds
 - it is intended to be reverted immediately
 - its contents are part of the main trunk, for future reference

Alefeld et al. demonstrate that TOMS 748 marginally outperforms Brent's
method on a particular suite of test problems. Brent's method is
nonetheless superior for the 388 solves in lmi's regression tests:

  evaluations  max   mean  sample-std-dev    commit        date
      7332      65   18.9       7.13       028b4541c  20210728T0052Z
      7622      76   19.6       7.65          this        today

This doesn't prove that either method is better in general. It does show
that neither is universally better than the other: practical outcomes
depend on the particular tests used.

To reproduce those tests, use the following trivial patch:

== not the usual scissors line, to avoid confusion ==
| --- a/ihs_avsolve.cpp
| +++ b/ihs_avsolve.cpp
| @@ -482,7 +482,8 @@ currency AccountValue::Solve
|          }
|
|      SolveHelper solve_helper(*this, solve_set_fn);
| -    root_type const solution = decimal_root
| +//  root_type const solution = decimal_root
| +    root_type const solution = toms748_root
|          (solve_helper
|          ,lower_bound
== not the usual scissors line, to avoid confusion ==

This command reproduces the 7332 total above without that patch:

  < /opt/lmi/test/trace.txt sed -e'/nominal/!d' -e's/ nom.*//' \
  -e's/^.*://' -e's/^ *[^ ]* //' | xargs | sed -e 's/\ /+/g' | bc

The 7622 total with the patch can be reproduced thus:

  < /opt/lmi/test/trace.txt sed -e'/evaluations/!d' \
  -e's/ *evaluations.*//' | xargs | sed -e 's/\ /+/g' | bc