Further improve a local function in a unit-test TU
The adjustments and supporting rationalizations in this and the
preceding commit suggest that it might be better to replace
max_n_eval_bolzano() completely. However, that would require a
considerable effort: the canonical ⌈log2((b-a)/tol)⌉ formula
disregards the actual value of the root, which Brent introduces
in his formula for bisection and then reuses in his formula for
the maximum number of iterations with his algorithm.