Fail concinnity check if there are any coding rules test failures
The changes of a9bc62f53 (Count coding-rules complaints, 2021-06-28)
made test_coding_rules exit with a non-zero exit code if any coding
rules violations were found, but the check_concinnity makefile target
still explicitly ignored the failure by using "-" in the make recipe.

It is not really clear why was the exit code of test_coding_rules
ignored in the first place, as it hadn't been done for the checks
performed in the makefile itself prior to the introduction of a separate
test_coding_rules program in 7afe844dc (Quadruple the speed of 'make
check_conformity', 2006-01-18), but it remained the case ever since.

Stop doing it now and do fail if either any real problems are found or
when running the stub version of test_coding_rules built on the systems
without PCRE.