Generally force group permissions to match user permissions
In a chroot that is to be shared by multiple developers using the same
$prefix, files in that $prefix should have the same permissions for the
developers' common group as for each user individually. Otherwise, some
files created by one user might not be modifiable by another. Generally
the permissions are already correct in a chroot created by lmi scripts,
but autotools-generated files are an exception because they don't heed
directories' GID bits.

It can of course be argued that each user should use its own prefix.