Fix LMI_WX_CHILD_DOCUMENT definition to fix g++11 C++20 build
Combining elements of two different enums, such as wxDOC_SILENT and
LMI_WX_CHILD_DOCUMENT, is not allowed in C++20 and g++11 warns about it.

Avoid this warning simply by not making LMI_WX_CHILD_DOCUMENT an enum
element at all but long, which is the type actually used by wxWidgets
for the "flags" parameter.