Fix gcc11 -Wdeprecated-enum-enum-conversion in test_coding_rules
Don't perform bit arithmetic operations between elements of different
enums, as this is not allowed in C++20 any longer.

Simply cast the enum value to its underlying type (i.e. int) to avoid
the warning, while still doing the same thing because we know that this
operation is safe in this particular case.