Make it simpler to search for errors in the CI build logs
Avoid using "***" in the workflow file itself, as otherwise searching
for these characters always found their occurrences in the script itself
first before finding it in the output.

Note that we still want to output "***" if an error really occurs, so
just break them into separate arguments, but still keep them.