Use modified argument instead of one of its addends
See:
  https://lists.nongnu.org/archive/html/lmi/2021-12/msg00001.html

This comment:
  /// Non-tiered complements are added to each argument as needed.
plainly seems to state that the intention was to treat all three
'double' arguments as increments to non-tiered values and use the
values thus incremented. This change has no effect on regression
testing, which proves only that there's no test for it; however, in
AccountValue::ApplyDynamicMandE(), the only place that function is
called, a nonzero value of the argument raises a
  "Tiered investment management fee unimplemented."
alarum, so any test for this feature would fail anyway.