Rework a redhat mount
Only comments have been changed. On a corporate server that uses lots
of little partitions for e.g. {/usr /home /opt /opt/nxlog ...}, the
uncomfortably small root partition overflowed because of the (former)
contents of /srv/cache_for_lmi . To avoid that recurring problem, moved
all of /srv only a comparatively gigantic 80 GiB partition. It is hoped
that redhat is rational enough not to depend on using /srv during the
boot process; encouragingly, /srv is empty except for lmi stuff.