Update libxml2 to the latest post v2.9.13 version
This fixes harmless but annoying warnings when running autogen.sh in
install_xml_libraries.sh and also includes a number of bug fixes,
including those fixing security problems, notably CVE-2021-3541 and
CVE-2022-23308, which don't directly affect lmi as it doesn't use
libxml2 to parse untrusted input, but are still worth including.

Note that the latest release (v2.9.13) doesn't include the changes to
configure.ac getting rid of the warnings, which is why we use the latest
master version.

Also note that the latest libxml2 installs documentation in a
version-independent directory, requiring a minor change to the
installation script.