Refresh lib{xml2,xslt} build files if necessary
If the build files already exist locally, rerun autoreconf to update
them if necessary, e.g. if the dependencies (such as configure.ac) or
the tools themselves changed.

This fixes build errors when building on a machine with an existing lmi
build after the recent libxslt submodule update.

This commit is best viewed ignoring whitespace-only changes.