Update wxPdfDoc to the latest version compatible with wx 3.2.0
The previously used version of wxPdfDoc couldn't be built with wx 3.1.6
or later, which exposed a latent problem in wxPdfDoc that was fixed in
the upsteam since then.

This updates also pulls some fixes from wxPdfDoc main branch (née
master), at least one of which affects lmi, see the upcoming commit.