Fix checkout step in the CI workflow to work with Git 2.35.2
This Git version adds a new check for insecure ancestor directories (see
CVE-2022-24765 for more details), which is triggered in the GitHub
Actions environment.

As there are no real security concerns in this particular case
(everything is entirely under our own control anyhow), just disable the
check for the directory which is used.