Avoid unnecessary rounding
Storing currency values as integer-valued doubles (cents) in the ledger
object has made rounding unnecessary in many cases, some of which were
dangerous. For example:

-    return round_specamt().c(InvariantValues().SpecAmt[year] / 100.0);
+    return from_cents(InvariantValues().SpecAmt[year]);

Suppose specamt is rounded up to the next whole dollar, which is known
to be the standard rule for at least one company. Then suppose the
result of the division above is 1234567 plus epsilon. Then the intended
value of 1234567 is incorrectly changed to 1234568. This commit removes
that potential flaw.