Clean up beforehand, not afterward
Formerly, files created as testing artifacts in a temporary directory
were obliterated at the end of the run. This made it less easy to
discern the reason for any failure. And anything left over from an
earlier run became a demonstration of Murphy's Law. And the cleanup
command was in effect
  rm "$throwaway_dir"/*
which could have a surprising effect under various circumstances.

Now, the temporary directory is obliterated before its first use, by
  rm -rf "$throwaway_dir"
which is less unsafe to copy and paste to the command line--so no old
artifacts can cause trouble, and the files remain for analysis after
the run completes (or fails to complete). This does waste two megabytes
of storage, but anyone who cares can clean up manually.