Store all "scalable" vectors in ledger as cents
"Scalable" columns are scaled so that the most extreme positive or
negative value fits without overflowing a format like "999,999,999".
The scaling factor naturally must apply to all columns, so that a
caption like "values in thousands" describes them all. Therefore,
the extrema must embrace all values in all these vectors, which is
easiest to accomplish when all use the same unit. Now they all use
the same unit, cents.

As an alternative, a multiplier (say, 1.00 for dollars and 100.0
for cents) could be stored somewhere and applied to the values for
determination of the extrema. That just seemed less simple.

As an alternative, the scaling factor could be based on some
presumptively maximal column like EOY death benefit. That would be
less robust, though: for example, some other column might exceed
-0.1 times the maximal death benefit, and the negative sign occupies
an extra character in the output; for another example, the salary
column is unbounded, so that a group death benefit of three times
salary to a maximum of $1M would produce a $1M death benefit for
an oligarch with a salary orders of magnitude higher.

The scaling logic pays little attention to formats that contain
decimals. Generally, that doesn't matter because such amounts are
tiny. The only one that seems relevant here is "AddonMonthlyFee":

     ,{"AddonCompOnPremium"              , "999,999,999"}
-    ,{"AddonMonthlyFee"                 , "999,999,999"}
+    ,{"AddonMonthlyFee"                 , " 999,999.99"}
     ,{"AnnGAIntRate_Current"            ,      "99.99%"}

which is surpassingly unlikely to be large: it looks weird because
the leftmost character is blank, but ",999,999.99" would clearly
be wrong.