Refactor for clarity
The present goal is to make all expressions involving
ldbl_eps_plus_one_times() as similar as possible.

'TgtPremMonthlyPolFee' is already a currency amount--an integral number
of cents--and (12/mode) is an integer, so their product is an integral
number of cents, and can therefore be hoisted out of the expression
that is be rounded.