Improve a debatable workaround
It seems unlikely that the situation described in commit f97235aed48a52
could be a problem for the inverse operation, i.e., calculating premium
according to a corridor strategy. For a 4.9000000000000003552713679
corridor factor that is notionally 490/100 but has representation error:
  specamt = roundup  ($100000 * factor) // 490001: wrong
because 0.00000000003552713679 rounds up to 1; but
  premium = rounddown($100000 / factor) // 20408.1632653...
has no such problem. Thus, this function's body might be written as:
  return round_max_premium().c((a_specamt / rate) / a_mode);

Instead, however, committed an implementation that closely parallels
BasicValues::GetModalSpecAmtCorridor(), lest a corridor factor like
250% be read from file as, say, 2.5 * (1 + epsilon), however unlikely
that may seem.

Made no such change in AccountValue::TxAscertainDesiredPayment(),
where a similar calculation is used for an error message. In that case
(an exotic idiosyncrasy of a particular product), the premium is
multiplied by 'MinSinglePremiumMult', whose value needn't be rational,
even notionally.