Refactor, almost
In life insurance, many rates are given in tables with a fixed number of
decimal places, and are stored as binary64 floating-point values because
that seems natural, yet are thought of as exact rational numbers whose
denominators are a small integral number of ten. Thus, in the example
discussed in commit f97235aed48a52, multiplying $100000.00 by 4.90 and
rounding up to the next dollar would give $490001 in floating-point
arithmetic, but that was considered glaringly wrong: instead, it is
considered exclusively correct to calculate 100000 * 490 / 100 in
rational integer arithmetic. Such cases abound.

Factored a new rate_times_currency() out of max_modal_premium() for
general use.

Reimplemented max_modal_premium() in terms of rate_times_currency().
This is a refactoring except in the case that the rate is given to more
than eight decimals, which is uncommon. In that case, no attempt is made
to recover the notional fixed-precision rational number that the rate is
taken to represent, and binary64 arithmetic is performed as formerly,
but with division by the modal factor {1, 2, 4, 12} applied after the
multiplication. This can make a difference in binary64 arithmetic, but
it makes the answer more correct for reasons explained inline in the
max_modal_premium() function. However, no regression is observed in
practice, so writing this paragraph was hardly worthwhile.