Use rate_times_currency()
Motivating example from an actual regression test:
  0.00000250 specified-amount load
     $250000 specified amount
In binary64 arithmetic, this is
  0.00000250 * 250000 = 0.625 (approximately!)
which is to be rounded to the nearest or even cent; 62 and 63 cents are
equally near, but 62 is even, so the answer should be 62.

However, for the i686+x87 architecture still used in production for now,
the answer was sixty-three cents, evidently because of representation
error in "0.00000250", which is eradicated by shifting the decimal point
rightward to produce a rational number with a power-of-ten denominator.

In the worst regression between i686 and x86_64, this difference of one
cent in a monthly deduction grew to $79.19 at compound interest over 75
years. Reasonableness check: the implicit annual percentage rate is
  9.04% = (79.19 / 0.12) ^ (1 / 75) - 1
which is acceptably close to the 7.46% separate-account rate illustrated
(the rates wouldn't be exactly the same because the account-value
calculation has many other components). That case represented the
largest discrepancy in absolute terms; with this commit, it matches on
both i686 and x86_64 to within 1e-12 dollars.

This is no mere pedantic quibble: US courts have entertained lawsuits
concerning the rounding of insurance rates. See, for example, the
following cases in the Zavala County Texas District Court:
 - 95-08-09169-CV Martinez v. Allstate Insurance Co.
 - 95-08-09165-CV Sendejo v. Texas Farmers Insurance Co.