Remove hard-coded wx{DatePicker,Spin}Ctrl sizes from XRC files
The sizes specified for wxSpinCtrl were too small and resulted in
truncating the control "-/+" buttons in the GTK-based version, as well
as tons of GTK debugging messages.

The size specified for wxDatePickerCtrl didn't result in any apparent
problems, but it could, potentially, depending on the theme and the date
format in the current locale, so don't hard-code the sizes for them
neither, as this is at best useless.

This generalizes the changes done in caee09caa (Remove invalid spinctrl
sizes, 2020-07-24) to all spin controls and wxDatePickerCtrl too.