Note an arcanum
There's a reason why this assertion has long been made for sales_load_:
    LMI_ASSERT(0.0 <= sales_load_);
but not for premium_load_. Writing one for premium_load_ and commenting
it out is less obscure than omitting it.

The reason is that a "discount" was to be given for a certain subplan to
reflect lower commissions. Notionally, the "sales load" represents
commissions, but in practice it was already less than the commission
reduction, so it was expeditious to set a negative rate for this load.