Prefer a calculation with less roundoff error
Hoist a constant 'k', using
  k * (r0 + r1 + ... + rN)
in preference to
  k * r0 + k * r1 + ... + k * rN
when they're materially the same, because the former is less susceptible
to roundoff, which differs between i686 and x86_64. Infrequently, one of
the terms in the summation is
  ... + k' * rm + ...
where "k'" differs from the usual 'k'; in that case, use the more
complicated formula despite its greater roundoff error.